Lab1 Control:
Some bugfixes, plus rudimentary demonstration.

While there are flaws in the design, for example, a terminate() cannot be followed by a spawn() in the same loop/ISR, which may be future problem to fix, overall with careful consideration, the design is very flexible and can create a wide variety of heirarchies correctly.

Created version @ 0.1.1